refactor(atomix): introduce PartitionDistributor
Extracts the PartitionDistributor interface and renames the previous
implementation to RoundRobinPartitionDistributor. This is a preparation
step to support different partitioning schemes, notably a manual one to
support multi-site configuration.

This commit also allows configuring the desired partition distributor
internally, though it defaults to the previous implementation.
