+++ atomix/cluster/src/main/java/io/atomix/raft/partition/RaftPartitionConfig.java
@@ -27,6 +27,8 @@ public class RaftPartitionConfig {
   private static final Duration DEFAULT_REQUEST_TIMEOUT = Duration.ofSeconds(5);
   private static final int DEFAULT_MIN_STEP_DOWN_FAILURE_COUNT = 3;
   private static final Duration DEFAULT_MAX_QUORUM_RESPONSE_TIMEOUT = Duration.ofSeconds(0);
+  private static final RoundRobinPartitionDistributor DEFAULT_PARTITION_DISTRIBUTOR =
+      new RoundRobinPartitionDistributor();
 
   private Duration electionTimeout = DEFAULT_ELECTION_TIMEOUT;
   private Duration heartbeatInterval = DEFAULT_HEARTBEAT_INTERVAL;
@@ -36,6 +38,7 @@ public class RaftPartitionConfig {
   private Duration requestTimeout = DEFAULT_REQUEST_TIMEOUT;
   private int minStepDownFailureCount = DEFAULT_MIN_STEP_DOWN_FAILURE_COUNT;
   private Duration maxQuorumResponseTimeout = DEFAULT_MAX_QUORUM_RESPONSE_TIMEOUT;
+  private PartitionDistributor partitionDistributor = DEFAULT_PARTITION_DISTRIBUTOR;
 
   /**
    * Returns the Raft leader election timeout.
@@ -145,5 +148,13 @@ public class RaftPartitionConfig {
    */
   public void setMaxQuorumResponseTimeout(final Duration maxQuorumResponseTimeout) {
     this.maxQuorumResponseTimeout = maxQuorumResponseTimeout;
+  }
+
+  public PartitionDistributor getPartitionDistributor() {
+    return partitionDistributor;
+  }
+
+  public void setPartitionDistributor(final PartitionDistributor partitionDistributor) {
+    this.partitionDistributor = partitionDistributor;
   }
 }
